Handle weights with dims >2.1B (#1819)
* Store weights dims as size_t

* Support size_t weight dims in unit tests

* Fix sign-compare error

+++ src/weights/weights.cpp
@@ -45,8 +45,8 @@ namespace {
  *  The tensor is stored in a matrix, although there may be multiple
  *  dimensions corresponding to the matrix height and width.
  */
-std::string get_dims_string(const std::vector<int>& matrix_height_dims,
-                            const std::vector<int>& matrix_width_dims) {
+std::string get_dims_string(const std::vector<size_t>& matrix_height_dims,
+                            const std::vector<size_t>& matrix_width_dims) {
   std::stringstream ss;
   ss << "(";
   for (size_t i = 0; i < matrix_height_dims.size(); ++i) {
@@ -122,35 +122,35 @@ description weights::get_description() const {
 // Dimension accessors
 // -----------------------------------------------
 
-std::vector<int> weights::get_dims() const {
-  std::vector<int> dims;
+std::vector<size_t> weights::get_dims() const {
+  std::vector<size_t> dims;
   for (const auto& d : get_matrix_width_dims())  { dims.push_back(d); }
   for (const auto& d : get_matrix_height_dims()) { dims.push_back(d); }
   return dims;
 }
-int weights::get_size() const {
+size_t weights::get_size() const {
   const auto& dims = get_dims();
   return std::accumulate(dims.begin(), dims.end(),
-                         1, std::multiplies<int>());
+                         1, std::multiplies<size_t>());
 }
-std::vector<int> weights::get_matrix_height_dims() const {
+std::vector<size_t> weights::get_matrix_height_dims() const {
   return m_matrix_height_dims;
 }
-std::vector<int> weights::get_matrix_width_dims() const {
+std::vector<size_t> weights::get_matrix_width_dims() const {
   return m_matrix_width_dims;
 }
-int weights::get_matrix_height() const {
+size_t weights::get_matrix_height() const {
   const auto& dims = get_matrix_height_dims();
   return std::accumulate(dims.begin(), dims.end(),
-                         1, std::multiplies<int>());
+                         1, std::multiplies<size_t>());
 }
-int weights::get_matrix_width() const {
+size_t weights::get_matrix_width() const {
   const auto& dims = get_matrix_width_dims();
   return std::accumulate(dims.begin(), dims.end(),
-                         1, std::multiplies<int>());
+                         1, std::multiplies<size_t>());
 }
-void weights::set_dims(std::vector<int> matrix_height_dims,
-                       std::vector<int> matrix_width_dims) {
+void weights::set_dims(std::vector<size_t> matrix_height_dims,
+                       std::vector<size_t> matrix_width_dims) {
   m_matrix_height_dims = std::move(matrix_height_dims);
   m_matrix_width_dims = std::move(matrix_width_dims);
   do_set_dims_(matrix_height_dims, matrix_width_dims);
@@ -204,7 +204,7 @@ void weights::setup_default_matrix_distribution() {
 void weights::setup() {
 
   // Check that tensor dimensions are valid
-  const auto& is_nonpositive = [] (int d) { return d <= 0; };
+  const auto& is_nonpositive = [] (size_t d) { return d <= 0; };
   if (std::any_of(m_matrix_height_dims.begin(),
                   m_matrix_height_dims.end(),
                   is_nonpositive)
